refactor: move I18n.js out of vendor/ (it's our own formatter now, not a dep)
Since conedevelopment/i18n was dropped, resources/js/vendor/I18n.js is no longer a
published third-party file — it's our own translation formatter. Move it to
resources/js/i18n/ and repoint the app.js + useTranslations imports.
